migrate TestExternalUserUpdateState to pytest
